+++ templates/admission/exports/includes/sic_refusal/REFUS_ARTICLE_95_SIC_CONDITIONS_PUBLIESS.html
@@ -37,9 +37,9 @@
 <p>
   Conformément à l’article 111, §1er, 4° du décret du 7 novembre 2013 définissant le paysage de l’enseignement
   supérieur et l’organisation académique des études, le jury impose des conditions de recevabilité. Celles-ci sont
-  précisées dans les conditions d'admission du programme visé (<a href="https://uclouvain.be/prog-2022-sinc1ba-cond_adm">
-  https://uclouvain.be/prog-2022-sinc1ba-cond_adm [En attente de développement pour la récupération du bon lien]</a>). Il
-  ressort de l'analyse de votre candidature que vous ne répondez pas à ces conditions.
+  précisées dans les conditions d'admission du programme visé
+  {% with conditions_url=proposition|access_conditions_url %}(<a href="{{ conditions_url }}">{{ conditions_url }}</a>)
+  {% endwith %}. Il ressort de l'analyse de votre candidature que vous ne répondez pas à ces conditions.
 </p>
 
 {% if proposition.motifs_refus %}
